fixed 64bit address issues and g++ compatibility related things

+++ Memory.hpp
@@ -3,8 +3,8 @@
 #include <Windows.h>
 #include <TlHelp32.h>
 #include <string>
-#include <psapi.h>
 #include <iostream>
+#include <psapi.h>
 #pragma comment(lib, "psapi")
 
 class Memory
@@ -14,11 +14,11 @@ public:
     long GetModuleBase(HANDLE processHandle, std::string &sModuleName);
     BOOL SetPrivilege(HANDLE hToken, LPCTSTR lpszPrivilege, BOOL bEnablePrivilege);
     BOOL GetDebugPrivileges(void);
-    int ReadInt(HANDLE processHandle, int address);
-    int GetPointerAddress(HANDLE processHandle, int startAddress, int offsets[], int offsetCount);
-    int ReadPointerInt(HANDLE processHandle, int startAddress, int offsets[], int offsetCount);
-    float ReadFloat(HANDLE processHandle, int address);
-    float ReadPointerFloat(HANDLE processHandle, int startAddress, int offsets[], int offsetCount);
-    char* ReadText(HANDLE processHandle, int address);
-    char* ReadPointerText(HANDLE processHandle, int startAddress, int offsets[], int offsetCount);
+    int ReadInt(HANDLE processHandle, long address);
+    int GetPointerAddress(HANDLE processHandle, long startAddress, int offsets[], int offsetCount);
+    int ReadPointerInt(HANDLE processHandle, long startAddress, int offsets[], int offsetCount);
+    float ReadFloat(HANDLE processHandle, long address);
+    float ReadPointerFloat(HANDLE processHandle, long startAddress, int offsets[], int offsetCount);
+    char* ReadText(HANDLE processHandle, long address);
+    char* ReadPointerText(HANDLE processHandle, long startAddress, int offsets[], int offsetCount);
 };
